requirements.txt
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.6
django-allauth==65.11.2
django-environ==0.12.0
et_xmlfile==2.0.0
git-filter-repo==2.47.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
pyzbar==0.1.9
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
